own project for EventBus, renamed game.events to game.communication and game.ticks
