+++ game/communication/src/main/kotlin/de/gleex/pltcmd/game/communication/TransmissionBuffer.kt
@@ -1,6 +1,6 @@
-package de.gleex.pltcmd.events.radio
+package de.gleex.pltcmd.game.communication
 
-import de.gleex.pltcmd.events.ticks.TickId
+import de.gleex.pltcmd.game.ticks.TickId
 import de.gleex.pltcmd.model.radio.communication.transmissions.Transmission
 import org.hexworks.cobalt.datatypes.Maybe
