+++ game/ticks/pom.xml
@@ -9,14 +9,9 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>game.radio-events</artifactId>
+    <artifactId>game.ticks</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>de.gleex.pltcmd</groupId>
-            <artifactId>model.radio</artifactId>
-            <version>${project.version}</version>
-        </dependency>
         <dependency>
             <groupId>de.gleex.pltcmd</groupId>
             <artifactId>util.events</artifactId>
