+++ util/pom.xml
@@ -13,7 +13,7 @@
     <packaging>pom</packaging>
     <modules>
         <module>tests</module>
-        <module>util.events</module>
+        <module>events</module>
     </modules>
 
 </project>
